chore(quaint): refactor tests for connectors, addressing feedback
