+++ quaint/src/connector/sqlite/params.rs
@@ -103,3 +103,29 @@ impl TryFrom<&str> for SqliteParams {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn sqlite_params_from_str_should_resolve_path_correctly_with_file_scheme() {
+        let path = "file:dev.db";
+        let params = SqliteParams::try_from(path).unwrap();
+        assert_eq!(params.file_path, "dev.db");
+    }
+
+    #[test]
+    fn sqlite_params_from_str_should_resolve_path_correctly_with_sqlite_scheme() {
+        let path = "sqlite:dev.db";
+        let params = SqliteParams::try_from(path).unwrap();
+        assert_eq!(params.file_path, "dev.db");
+    }
+
+    #[test]
+    fn sqlite_params_from_str_should_resolve_path_correctly_with_no_scheme() {
+        let path = "dev.db";
+        let params = SqliteParams::try_from(path).unwrap();
+        assert_eq!(params.file_path, "dev.db");
+    }
+}
